renamed crypto files to make clear they are in the same class
