+++ src/crypto-lcg.h
